requirements.txt
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.9.0
fonttools==4.49.0
fsspec==2024.2.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightning-utilities==0.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pymystem3==0.2.0
pyparsing==3.1.1
pyproject-toml==0.0.10
python-dateutil==2.8.2
pytz==2024.1

PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
smart-open==7.0.1
smmap==5.0.1
stack-data==0.6.3
stqdm==0.0.5
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.20
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
wheel==0.42.0
wrapt==1.16.0
zipp==3.17.0